pyproject.toml
[project]
name = "python-oneshots"
version = "0.1.0"
description = "Mini scripts I've used as exercises and tools for my games. I hope they help you too."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "jsonschema>=4.26.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"